pyproject.toml
[tool.poetry]
name = "interview"
version = "0.1.0"
description = "hey ! :))"
authors = ["John Proud"]

[tool.poetry.dependencies]
python = "^3.9"
alembic = "^1.6.5"
Flask-SQLAlchemy = "^2.5.1"
flask-marshmallow = "^0.14.0"
Flask-RESTful = "^0.3.9"
Flask-Migrate = "^3.0.1"
psycopg2-binary = "^2.9.1"
sqlalchemy-mixins = "^1.4"
Flask-Cors = "^3.0.10"
openpyxl = "^3.0.7"
python-dateutil = "^2.8.1"
python-dotenv = "^0.17.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
